### 2016-03-11 #29 @njahn82
rebi - an R-Client for Europe PMC
* 1. What does this package do? (explain in 50 words or less)

rebi gives access to Europe PubMed Central, an indexing service for life-
science publications that is provided by the European Bioinformatics Institute
(EBI). This client can be used to search metadata and full-texts, retrieve
reference sections, citations, text-mined terms, and links to other EBI
databases or external sources like the Wikipedia.

* 2. Paste the full DESCRIPTION file inside a code block below.

```
Package: rebi
Version: 0.0.999
Maintainer: 'Najko Jahn' <najko.jahn@uni-bielefeld.de>
Author: Najko Jahn
Title: rebi -- R Interface for Europe PMC RESTful Web Service
Description: R Interface to Europe PMC RESTful Web Service. Europe PMC covers
    life science literature and it gives access to open access full texts.
    Coverage is not only restricted to Europe, but articles and abstracts are
    indexed from all over the world. As a partner in the PMC International
    (PMCi), Europe PMC ingests all PubMed content and extends its index with
    other sources, including Agricola, a bibliographic database of citations to
    the agricultural literature, or Biological Patents. In addition to
    bibliographic data, rebi gives automated access to citations and references
    that were indexed by Europe PMC. Links between life-science literature and
    other EBI databases, including ENA, PDB or ChEMBL are also accessible.
    External links provide a mechanism to link out to external providers such as
    Wikipedia or research data repositories.
LazyLoad: yes
LazyData: yes
    httr,
    plyr,
    xml2

```

* 3. URL for the package (the development repository, not a stylized html page)

https://github.com/njahn82/rebi

* 4. What data source(s) does it work with (if applicable)?

It only works with Europe PMC's Articles RESTful API.
https://europepmc.org/RestfulWebService

* 5. Who is the target audience?

Life-science researchers and students, scholars that study the life sciences,
librarians.

* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?

To my knowledge, this is the only package that implements Europe PMC's
Articles RESTful API. Since there is a considerable overlap with PubMed/PubMed
Central, the `rentrez` package could be used as well to fetch bibliographic
information. `fulltext` package gives access to supplementary material
deposited in Europe PMC. `oai` package could be used to retrieve metadata from
Europe PMC via its OAI-PMH interface.

* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
    * [X] This package does not violate the Terms of Service of any service it interacts with.
    * [X] The repository has continuous integration with Travis CI and/or another service
    * [X] The package contains a vignette
    * [X] The package contains a reasonably complete README with `devtools` install instructions
    * [X] The package contains unit tests
    * [X] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
    * [ ] Are there any package dependencies not on CRAN?
    * [X] Does the package have a CRAN accepted license?
    * [x] Did `devtools::check()` produce any errors or warnings? If so paste them below.
* 9. Please add explanations below for any exceptions to the above:
* 10. If this is a resubmission following rejection, please explain the change in circumstances.

This package has been available through rOpenSci since summer 2013. Because I
made major upgrades in the last days that reflect the rOpenSci guidance on how
to write packages, I thought it might be helpful to re-submit this package.

See also the discussion in the forum:
https://discuss.ropensci.org/t/major-package-update-of-rebi-and-question-regarding-re-submission/333/
