### 2016-03-28 #29 @toph-allen
Here's my review. N.B. According to the submission above, the package has been available through rOpenSci since 2013, but has recently made major upgrades that reflect rOpenSci guidance on package authorship, so it probably is already sufficient.  Let me know if you have any questions, or any areas where you'd like me to clarify more.

## Installation

- No issues installing on OS X.

## General comments on code

- All sample code ran correctly.
- The code generally seems concise and sensibly divided into different functions. As far as I can tell, the package addresses the problem space pretty comprehensively, and the code is clear to understand and pretty self-explanatory. Nice work. 👍

### Comments on specific functions

- `epmc_ftxt()` includes a `data_src = "pmc"` argument, but the function is hard-coded to always use PMC as a source--this argument is only used to throw an error if anything other than PMC is used. Perhaps better to remove this argument and clarify in documentation, as using `args(epmc_ftxt)` might sugget to a user that they *can* use another source.

### Improvements to code style?

- There are a few local package variables in use: `supported_dbs` and `supported_semantic_types` are used in one function each, and are defined in that function's file; `supported_data_src` is used in many functions, but defined in `epmc_details.r`. I'm not sure if there's an accepted best practice for this, but it seems it'd make more sense to either put these inside functions or put them in their own file, say `supported.r`.
- While code itself is self-explanatory enough that I don't think this is a major issue, sometimes comments are a little sparse or terse (i.e. just a word or two).

### Comments on documentation?

- The top-level package help file is `rebi-package.r`. Seems like it would be better to have this as `rebi.r` and change `@name rebi-package` to `@name rebi`, so that `?rebi` brings up that document.
- There are one or two places where documentation could use a touch of clarifying.
  - E.g. `epmc_citations.r` could be clearer as to whether it fetches works *cited by* a publication vs. works *citing* a publication.
  - `epmc_refs.r` has the same problem, especially because it refers to "citations" in its documentation. This could be simply a reflection of the language in the underlying API's documentation, but I think it could use a little clarification.

### Do tests pass locally?

- Yes

### Documentation and examples run without issue

- Yes

## ROpenSci Packaging Guidelines

- Package naming
  - It's a fine name, though I don't know its provenance.
- Function/variable naming
  - Snake case, self-explanatory. All functions are prefixed with "epmc_", which seems unnecessary? It seems, however, that this at least in part to avoid naming conflicts with defunct functions. The new naming scheme, while a bit redundant, seems more consistent than the old one.
- README
  - README.md is clear and very comprehensive, explaining what the package can do and how to do it.
- Code of conduct
  - Yep, included
- Documentation
  - Uses `roxygen2`
- Package dependencies
- Testing
  - Uses `testthat`
- Continuous integration
  - Yep, TravisCI
- Console messages
  - No instances of `cat()` or `print()` in the package
- Recommended software scaffolding
  - Uses all recommended packages where relevant

(I spent about 1.5–2 hours on the review).
