### 2016-03-28 #29 @noamross
Thanks for the review @toph-allen!

@njahn82 Let us know when you've incorporated / responded to @toph-allen's comments and we can merge the updated version of **rebi** into the rOpenSci repo.
