### 2016-04-15 njahn82/rebi#2 @njahn82
remove data_src parameter in epmc_ftxt()
From the onboarding review:

> `epmc_ftxt()` includes a `data_src = "pmc"` argument, but the function is hard-coded to always use PMC as a source--this argument is only used to throw an error if anything other than PMC is used. Perhaps better to remove this argument and clarify in documentation, as using `args(epmc_ftxt)`` might sugget to a user that they can use another source.
