### 2016-04-15 caa7479
removed . Instead, it checks if a valid PMCID is provided. closes #2
