+++ man/epmc_tm_count.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/empc_tm_count.r
+% Please edit documentation in R/epmc_tm_count.r
 \name{epmc_tm_count}
 \alias{epmc_tm_count}
 \title{Retrieve the number of database links from Europe PMC publication database}
