### 2016-04-15 ffe8ed6
move commonly used functions including uri and path variables to utils.r #3
