+++ R/epmc_db_count.r
@@ -45,7 +45,7 @@ epmc_db_count <- function(ext_id = NULL, data_src = "med"){
                 following sources: ", paste0(supported_data_src, collapse =", ")
     ))
   # build request
-  path = paste("europepmc/webservices/rest", data_src, ext_id, "databaseLinks",
+  path = paste(rest_path(), data_src, ext_id, "databaseLinks",
                "/json", sep ="/")
   doc <- rebi_GET(path = path)
   if(is.null(doc$dbCountList))
