+++ R/epmc_ftxt.r
@@ -20,9 +20,8 @@ epmc_ftxt <- function(ext_id = NULL) {
   if (!substring(ext_id, 1, 3) == "PMC")
     stop("Full texts can only be provided for publications indexed PubMed
          Central. So please provide a PMCID, i.e. ids starting with 'PMC'")
-  uri <- "http://www.ebi.ac.uk"
   # call api
-  req <- httr::GET(uri, path = paste("europepmc/webservices/rest", ext_id,
+  req <- httr::GET(base_uri(), path = paste(rest_path(), ext_id,
                                      "fullTextXML", sep = "/"))
   # check for http status
   httr::stop_for_status(req, "retrieve full text. Check with epmc_details()
