+++ R/epmc_tm_count.r
@@ -32,7 +32,7 @@ epmc_tm_count <- function(ext_id = NULL, data_src = "med"){
                 following sources: ", paste0(supported_data_src, collapse =", ")
     ))
   # build request
-  path = paste("europepmc/webservices/rest", data_src, ext_id, "textMinedTerms",
+  path = paste(rest_path(), data_src, ext_id, "textMinedTerms",
                "/json", sep ="/")
   doc <- rebi_GET(path = path)
   if(doc$hitCount == 0)
