+++ R/epmc_search.r
@@ -33,7 +33,7 @@ epmc_search <- function(query = NULL, id_list = FALSE, n_pages = 50){
     stop("No query provided")
   if (!is.numeric(n_pages))
     stop("n_pages must be of type 'numeric'")
-  path = "europepmc/webservices/rest/search"
+  path = paste0(rest_path(), "/search")
   q <- list(query = query, format = "json")
   doc <- rebi_GET(path = path, query = q)
   hitCount <- doc$hitCount
@@ -63,40 +63,4 @@ epmc_search <- function(query = NULL, id_list = FALSE, n_pages = 50){
   md <- result[, !(names(result) %in% fix_list(result))]
   # return
   list(hit_count = hitCount, data = md)
-}
-
-# Implementing GET method and json parser for EPMC
-rebi_GET <- function(path = NULL, query = NULL, ...) {
-  if (is.null(path) && is.null(query))
-    stop("Nothing to search")
-  uri <- "http://www.ebi.ac.uk"
-  # call api
-  req <- httr::GET(uri, path = path, query = query)
-  # check for http status
-  httr::stop_for_status(req)
-  # load json into r
-  out <- httr::content(req, "text")
-  # valid json
-  if(!jsonlite::validate(out))
-    stop("Upps, nothing to parse, please check your query")
-  doc <- jsonlite::fromJSON(out)
-  if (!exists("doc"))
-    stop("No json to parse", call. = FALSE)
-  doc
-}
-
-# Calculate pages. Each page consists of 25 records.
-rebi_pageing <- function(hitCount, pageSize) {
-  if (all.equal((hitCount / pageSize), as.integer(hitCount / pageSize)) == TRUE) {
-    1:(hitCount / pageSize)
-  } else {
-    1:(hitCount / pageSize + 1)
-  }
-}
-
-# fix to remove columns that cannot be easily flatten from the data.frame
- fix_list <- function(x){
-   if(!is.null(x))
-  tmp <- plyr::ldply(x, is.list)
-  tmp[tmp$V1 == TRUE, ".id"]
 }

[…]
