### 2016-04-15 6f97556
moved package documentation
