+++ man/rebi.Rd
@@ -1,7 +1,7 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rebi-package.r
-\name{rebi-package}
-\alias{rebi-package}
+% Please edit documentation in R/rebi.r
+\name{rebi}
+\alias{rebi}
 \title{rebi - R client for the Europe PMC RESTful article API}
 \description{
 What is rebi?
