+++ R/epmc_refs.r
@@ -50,7 +50,7 @@ epmc_refs <- function(ext_id = NULL, data_src = "med", n_pages = 20) {
   if(max(no_pages) > n_pages) no_pages <- 1:n_pages
   pages = list()
   for(i in no_pages){
-    out <- rebi_GET(path = paste("europepmc/webservices/rest", data_src, ext_id,
+    out <- rebi_GET(path = paste(rest_path(), data_src, ext_id,
                                  "references", "json", i, sep ="/"))
     message("Retrieving page ", i)
     result <- plyr::ldply(out$referenceList, data.frame,
