### 2016-04-16 4b82692
improved control of search limits
