+++ R/epmc_search.r
@@ -9,8 +9,11 @@
 #'   build a search query, see \url{http://europepmc.org/Help}
 #' @param id_list Should only IDs (e.g. PMID) and sources be retrieved for the
 #'   given search terms?
-#' @param n_pages Number of pages to be returned. By default, this function
-#'   returns 25 records for each page.
+#' @param batch_size Size of pages to be returned. By default, this function
+#'   returns 100  records for each page.
+#' @param limit Maximum number of recors to be retrieved. By default, 100 are
+#'   returned.
+#' @param verbose	print some information on what is going on.
 #' @return List of two, number of hits and the retrieved metadata as data.frame
 #' @examples \dontrun{
 #' #Search articles for 'Gabi-Kat'
@@ -24,43 +27,76 @@
 #'
 #' #Get only PLOS Genetics article with EMBL database references
 #' my.data <- epmc_search(query = 'ISSN:1553-7404 HAS_EMBL:y')
+#' #Limit search to 250 PLOS Genetics articles
+#' my.data <- epmc_search(query = 'ISSN:1553-7404', limit = 250)
+#'
 #' }
 #' @export
 
-epmc_search <- function(query = NULL, id_list = FALSE, n_pages = 50){
+epmc_search <- function(query = NULL, batch_size = 100, limit = 1000,
+                         id_list = FALSE, verbose = TRUE){
   # check
   if (is.null(query))
     stop("No query provided")
-  if (!is.numeric(n_pages))
-    stop("n_pages must be of type 'numeric'")
+  # get results found
   path = paste0(rest_path(), "/search")
   q <- list(query = query, format = "json")
   doc <- rebi_GET(path = path, query = q)
-  hitCount <- doc$hitCount
-  if(hitCount == 0)
+  hit_count <- doc$hitCount
+  if(hit_count == 0)
     stop("nothing found, please check your query")
-  no_pages <- rebi_pageing(hitCount = hitCount, pageSize = doc$request$pageSize)
-  # limit number of pages that will be retrieved
-  if(max(no_pages) > n_pages) no_pages <- 1:n_pages
-  pages = list()
-  for(i in no_pages){
-    if(!id_list) {
-      out <- rebi_GET(path = path,
-                      query = list(query = query,format = "json", page = i))
-      } else {
-        out <- rebi_GET(path = path,
-                        query = list(query = query, format = "json", page = i,
-                                     resulttype ="idlist"))
-      }
-    message("Retrieving page ", i)
-    result <- plyr::ldply(out$resultList, data.frame,
-                          stringsAsFactors = FALSE, .id = NULL)
-    pages[[i+1]] <- result
+  #prepare queries
+  limit <- ifelse(hit_count <= limit, hit_count, limit)
+  if (limit >= batch_size) {
+    if (all.equal(limit / batch_size, as.integer(limit / batch_size)) == TRUE) {
+      page_max <- limit / batch_size
+      last_chunk <- batch_size
+    } else {
+      page_max <- as.integer(limit / batch_size) + 1
+      last_chunk <- limit - ((page_max - 1) * batch_size)
     }
+    queries <-
+      lapply(1:(page_max - 1),
+             build_query,
+             batch_size = batch_size,
+             query = query)
+    queries <-
+      append(queries, list(
+        build_query(
+          query = query,
+          page = page_max,
+          batch_size = last_chunk
+        )
+      ))
+  } else {
+    queries <-
+      list(build_query(
+        page = 1,
+        query = query,
+        batch_size = limit
+      ))
+  }
+  # if only ids are requested
+  if (id_list == TRUE)
+    queries <-
+    lapply(queries, function(x)
+      c(x, resulttype = "idlist"))
[…]
