+++ tests/testthat/test_epmc_search.r
@@ -6,7 +6,7 @@ test_that("epmc_search returns", {
   b <- epmc_search(query = 'DOI:10.1007/bf00197367')
   c <- epmc_search(query = 'EXT_ID:22246381')
   d <- epmc_search(query = 'ISSN:1553-7404 HAS_EMBL:y')
-  e <- epmc_search(query = 'ISSN:1553-7404', id_list= TRUE, n_pages = 2)
+  e <- epmc_search(query = 'ISSN:1553-7404', id_list= TRUE, limit = 250)
 
   #correct class
   expect_output(str(a), "List of 2")
@@ -23,10 +23,9 @@ test_that("epmc_search returns", {
   expect_is(e$data, "data.frame")
 
   #are diminsions correct?
-  expect_equal(nrow(e$data), 50)
+  expect_equal(nrow(e$data), 250)
   expect_equal(ncol(e$data), 4)
 
   # fails correctly
   expect_error(epmc_search("123haha"), "nothing found, please check your query")
-  expect_error(epmc_search(query = 'ISSN:1553-7404', n_pages = "abc"))
 })
