+++ vignettes/rebi-vignettes.Rmd
@@ -87,7 +87,7 @@ head(my_data$data)
 Get list of ids that represent articles referencing DataCite DOIs:
 
 ```{r}
-my_data <- epmc_search(query = "ACCESSION_TYPE:doi", n_pages = 1, 
+my_data <- epmc_search(query = "ACCESSION_TYPE:doi", limit = 100, 
                        id_list = TRUE)
 head(my_data$data)
 my_data$hit_count
@@ -96,7 +96,7 @@ my_data$hit_count
 Use [ORCID](http://orcid.org/) to search for personal publications:
 
 ```{r}
-my_data <- epmc_search(query = 'AUTHORID:"0000-0002-7635-3473"', n_pages = 1) 
+my_data <- epmc_search(query = 'AUTHORID:"0000-0002-7635-3473"') 
 head(my_data$data)
 my_data$hit_count
 ```
