### 2016-04-19 d8f1c69
added search count function and epmc_search now returns a data.frame #4 ; set batch_size to 1000 and limit to 25 #5
