+++ R/epmc_search.r
@@ -9,10 +9,8 @@
 #'   build a search query, see \url{http://europepmc.org/Help}
 #' @param id_list Should only IDs (e.g. PMID) and sources be retrieved for the
 #'   given search terms?
-#' @param batch_size Size of pages to be returned. By default, this function
-#'   returns 100  records for each page.
-#' @param limit Maximum number of recors to be retrieved. By default, 100 are
-#'   returned.
+#' @param limit limit the number of records you wish to retrieve.
+#'   By default, 25 are returned.
 #' @param verbose	print some information on what is going on.
 #' @return List of two, number of hits and the retrieved metadata as data.frame
 #' @examples \dontrun{
@@ -33,32 +31,27 @@
 #' }
 #' @export
 
-epmc_search <- function(query = NULL, batch_size = 100, limit = 1000,
-                         id_list = FALSE, verbose = TRUE){
+epmc_search <- function(query = NULL, limit = 25, id_list = FALSE,
+                        verbose = TRUE){
   # check
   if (is.null(query))
     stop("No query provided")
   # get results found
-  path = paste0(rest_path(), "/search")
-  q <- list(query = query, format = "json")
-  doc <- rebi_GET(path = path, query = q)
-  hit_count <- doc$hitCount
-  if(hit_count == 0)
-    stop("nothing found, please check your query")
+  hit_count <- epmc_hits(query = query)
   #prepare queries
   limit <- ifelse(hit_count <= limit, hit_count, limit)
-  if (limit >= batch_size) {
-    if (all.equal(limit / batch_size, as.integer(limit / batch_size)) == TRUE) {
-      page_max <- limit / batch_size
-      last_chunk <- batch_size
+  if (limit > batch_size()) {
+    if (all.equal(limit / batch_size(), as.integer(limit / batch_size())) == TRUE) {
+      page_max <- limit / batch_size()
+      last_chunk <- batch_size()
     } else {
-      page_max <- as.integer(limit / batch_size) + 1
-      last_chunk <- limit - ((page_max - 1) * batch_size)
+      page_max <- as.integer(limit / batch_size()) + 1
+      last_chunk <- limit - ((page_max - 1) * batch_size())
     }
     queries <-
       lapply(1:(page_max - 1),
              build_query,
-             batch_size = batch_size,
+             batch_size = batch_size(),
              query = query)
     queries <-
       append(queries, list(
@@ -86,8 +79,7 @@ epmc_search <- function(query = NULL, batch_size = 100, limit = 1000,
     if(verbose == TRUE)
       message(paste0(hit_count, " records found. Retrieving batch ",
                      x[["page"]]))
-    out <-
-      rebi_GET(path = "europepmc/webservices/rest/search", query = x)
+    out <- rebi_GET(path = paste0(rest_path(), "/search"), query = x)
     plyr::ldply(out$resultList,
                 data.frame,
                 stringsAsFactors = FALSE,
@@ -97,6 +89,7 @@ epmc_search <- function(query = NULL, batch_size = 100, limit = 1000,
   result <- jsonlite::rbind.pages(out)
   # remove nested lists from data.frame, get these infos with epmc_details
   md <- result[, !(names(result) %in% fix_list(result))]
-  # return
-  list(hit_count = hit_count, data = md)
+  # return (thanks to @cstubben)
+  attr(md, "hit_count") <- hit_count
+  md
 }
