+++ vignettes/rebi-vignettes.Rmd
@@ -59,9 +59,8 @@ through other `rebi`-functions. Columns inform whether open access full texts
 text-mined terms (`hasTextMinedTerms`) or references (`hasReferences`) are
 available.
 
-By default, `epmc_search` returns 1.000 records as a maximum. This equals 50
-result pages that are returned by the API. To adjust it, simply use the
-`n_pages` parameter.
+By default, `epmc_search` returns 1.000 records. To adjust the limit, simply use 
+the `limit` parameter.
 
 Either list of publication ids (`id_list = TRUE`) or key metadata
 information  (`id_list = FALSE`, default option) are returned.
@@ -71,35 +70,37 @@ For instance, search for abstracts and full texts that mention `Gabi-Kat`:
 ```{r}
 library(rebi)
 my_data <- epmc_search(query = 'Gabi-Kat')
-# number of records
-my_data$hit_count
 # first six records
-head(my_data$data)
+head(my_data)
 ```
 
 Get PLOS Genetics (ISSN:1553-7404) articles that cross-reference EMBL:
 
 ```{r}
 my_data <- epmc_search(query = 'ISSN:1553-7404 HAS_EMBL:y')
-head(my_data$data)
+head(my_data)
 ```
 
-Get list of ids that represent articles referencing DataCite DOIs:
+Count hits before with `epmc_hits` to define limit. For example, get list of ids
+that represent articles referencing DataCite DOIs:
 
 ```{r}
-my_data <- epmc_search(query = "ACCESSION_TYPE:doi", limit = 100, 
+query <- "ACCESSION_TYPE:doi"
+epmc_hits(query)
+# set limit to 10 records
+my_data <- epmc_search(query = "ACCESSION_TYPE:doi", limit = 10, 
                        id_list = TRUE)
-head(my_data$data)
-my_data$hit_count
+head(my_data)
+attr(my_data, "hit_count")
 ```
 
 Use [ORCID](http://orcid.org/) to search for personal publications:
 
 ```{r}
 my_data <- epmc_search(query = 'AUTHORID:"0000-0002-7635-3473"') 
-head(my_data$data)
-my_data$hit_count
+attr(my_data, "hit_count")
 ```
+
 
 ## Get article details
 

[…]
