### 2016-04-29 b9cc515
better input param handling
