### 2016-04-29 15bc17f
generic fxns for limiting results to be returnded #5
