+++ tests/testthat/test_epmc_refs.r
@@ -4,7 +4,7 @@ test_that("epmc_refs returns", {
   skip_on_cran()
   a <- epmc_refs("PMC3166943", data_src = "pmc")
   b <- epmc_refs("25378340")
-  c <- epmc_refs("2439888", n_pages = 4)
+  c <- epmc_refs("2439888", limit = 10)
 
   #correct dimensions and class
   expect_output(str(a), "List of 2")
@@ -17,7 +17,7 @@ test_that("epmc_refs returns", {
   expect_is(c$data, "data.frame")
   expect_is(a$hit_count, "integer")
 
-  expect_equal(nrow(c$data), 100)
+  expect_equal(nrow(c$data), 10)
 
   # fails correctly
   expect_error(epmc_refs("14756321"), "No references found")
