### 2016-05-03 aa9860f
return hit_counts as attribute (#4) and make the distinction between epmc_refs and epmc_citations clearer
