+++ tests/testthat/test_epmc_citations.r
@@ -8,17 +8,15 @@ test_that("epmc_citations returns", {
   d <- epmc_citations("7535888", limit = 25)
 
   #correct dimensions and class
-  expect_output(str(a), "List of 2")
-  expect_output(str(b), "List of 2")
-  expect_output(str(c), "List of 2")
-  expect_output(str(d), "List of 2")
+  expect_output(str(a), "data.frame")
+  expect_output(str(b), "data.frame")
+  expect_output(str(c), "data.frame")
+  expect_output(str(d), "data.frame")
 
-  #correct class metadata
-  expect_is(a$data, "data.frame")
-  expect_is(b$data, "data.frame")
-  expect_is(c$data, "data.frame")
-  expect_is(d$data, "data.frame")
-  expect_is(a$hit_count, "integer")
+  expect_is(attr(a, "hit_count"), "integer")
+  expect_is(attr(b, "hit_count"), "integer")
+  expect_is(attr(c, "hit_count"), "integer")
+  expect_is(attr(d, "hit_count"), "integer")
 
   # fails correctly
   expect_error(epmc_citations("13814508"), "No citing documents found")
