+++ tests/testthat/test_epmc_refs.r
@@ -7,17 +7,15 @@ test_that("epmc_refs returns", {
   c <- epmc_refs("2439888", limit = 10)
 
   #correct dimensions and class
-  expect_output(str(a), "List of 2")
-  expect_output(str(b), "List of 2")
-  expect_output(str(c), "List of 2")
+  expect_output(str(a), "data.frame")
+  expect_output(str(b), "data.frame")
+  expect_output(str(c), "data.frame")
 
-  #correct class metadata
-  expect_is(a$data, "data.frame")
-  expect_is(b$data, "data.frame")
-  expect_is(c$data, "data.frame")
-  expect_is(a$hit_count, "integer")
+  expect_is(attr(a, "hit_count"), "integer")
+  expect_is(attr(b, "hit_count"), "integer")
+  expect_is(attr(c, "hit_count"), "integer")
 
-  expect_equal(nrow(c$data), 10)
+  expect_equal(nrow(c), 10)
 
   # fails correctly
   expect_error(epmc_refs("14756321"), "No references found")
