### 2016-05-03 #29 @njahn82
Hi  @toph-allen, sorry for coming back to you after such a long time. Thank you again for your helpful comments and suggestions to improve my package. I have a new version ready that include your remarks. I'll go over them in the following.

Please let me know if you have any further comments.

> epmc_ftxt() includes a data_src = "pmc" argument, but the function is hard-coded to always use PMC as a source--this argument is only used to throw an error if anything other than PMC is used. Perhaps better to remove this argument and clarify in documentation, as using args(epmc_ftxt) might sugget to a user that they can use another source.

As suggested, I removed the `data_src = "pmc"` argument from the function `epmc_ftxt()`

https://github.com/njahn82/rebi/issues/2

> There are a few local package variables in use: supported_dbs and supported_semantic_types are used in one function each, and are defined in that function's file; supported_data_src is used in many functions, but defined in epmc_details.r. I'm not sure if there's an accepted best practice for this, but it seems it'd make more sense to either put these inside functions or put them in their own file, say supported.r.

Agreed. I moved functions that are commonly used to `utils.r`. Thanks to your suggestion, I also started to dry out my code by

- defining the uri and the REST API path in utils.r only,
- use of roxygen's `@inheritParams`
- Thanks to @stubben , the default limit is now 25. Batch_size is by default 1000 and is dropped as function parameter. The function `make_queries` in `utils.r` implements paging and is used by `epmc_search()`, `make_path` provides a more generic way to generate paths for `epmc_ref()`, `epmc_citations()`, `epmc_db()`, `epmc_tm()`

> While code itself is self-explanatory enough that I don't think this is a major issue, sometimes comments are a little sparse or terse (i.e. just a word or two).

Agreed. Added more comprehensive explanations in some places.

> The top-level package help file is rebi-package.r. Seems like it would be better to have this as rebi.r and change @name rebi-package to @name rebi, so that ?rebi brings up that document.

I moved `rebi-package.r` to `rebi.r`

>  There are one or two places where documentation could use a touch of clarifying. E.g. epmc_citations.r could be clearer as to whether it fetches works cited by a publication vs. works citing a publication.

As suggested, I clarified both functions. In the documentation of `epmc_refs()` it says now:

```
#'  Get references for a given publication
#'
#'  This function retrieves all the works listed in the bibliography of a given
#'  article.
```

`epmc_citations()`

```
#' Get citations for a given publication
#'
#' The function finds works that cite a given publication.
```

> Package naming: It's a fine name, though I don't know its provenance.

It comes from the acronym European Bioinformatics Institute (EBI), which hosts Europe PMC. Do you think `europepmc` as package name would be more appropriate since the EBI offers many other services that are not covered by this package?

> Function/variable naming All functions are prefixed with "epmc_", which seems unnecessary?

I followed the example of other rOpenSci packages like `rcrossref` or `aRxiv`. Is there any best practise available if and when to prefix functions?

Again, many thanks for your helpful comments and suggestions.
