### 2016-05-10 #29 @noamross
Quick editor's comment about function naming: We recommend prefixing (e.g, `epmc_`) to prevent conflicts and ease autocompletion when function names are likely to conflict with other packages.  In this case, this is certainly the case, as `search()`, `ref()`, etc., will conflict with other packages and even the base R namespace.
