### 2016-05-13 90a4c17
allow to extend search with mesh and uniprot synonyms
