+++ README.md
@@ -110,7 +110,7 @@ head(my_data)
 #> 2                 Y                    N            N           N
 #> 3                 Y                    N            N           N
 #> 4                 Y                    N            N           N
-#> 5                 Y                    N            N           N
+#> 5                 Y                    Y            N           N
 #> 6                 N                    N            N           N
 #>   hasTMAccessionNumbers luceneScore issue
 #> 1                     N         NaN  <NA>
@@ -130,59 +130,59 @@ head(my_data)
 #>         id source     pmid      pmcid                          doi
 #> 1 26982327    MED 26982327 PMC4794157 10.1371/journal.pgen.1005920
 #> 2 27082250    MED 27082250 PMC4833346 10.1371/journal.pgen.1005954
-#> 3 26495848    MED 26495848 PMC4619825 10.1371/journal.pgen.1005609
-#> 4 26020649    MED 26020649 PMC4447368 10.1371/journal.pgen.1005207
-#> 5 25875245    MED 25875245 PMC4395465 10.1371/journal.pgen.1005124
-#> 6 25622028    MED 25622028 PMC4314295 10.1371/journal.pgen.1004970
-#>                                                                                                                                                     title
-#> 1                                                             Hybrid Dysgenesis in Drosophila simulans Associated with a Rapid Invasion of the P-Element.
-#> 2                                         Chromosomal-Level Assembly of the Asian Seabass Genome Using Long Sequence Reads and Multi-layered Scaffolding.
-#> 3                                                                                                     Virus Satellites Drive Viral Evolution and Ecology.
-#> 4                       A Simple Auxin Transcriptional Response System Regulates Multiple Morphogenetic Processes in the Liverwort Marchantia polymorpha.
-#> 5 MAPK signaling pathway alters expression of midgut ALP and ABCC genes and causes resistance to Bacillus thuringiensis Cry1Ac toxin in diamondback moth.
-#> 6                                                   Naturally occurring differences in CENH3 affect chromosome segregation in zygotic mitosis of hybrids.
+#> 3 27120580    MED 27120580       <NA> 10.1371/journal.pgen.1005987
+#> 4 27149082    MED 27149082       <NA> 10.1371/journal.pgen.1006030
+#> 5 26495848    MED 26495848 PMC4619825 10.1371/journal.pgen.1005609
+#> 6 26020649    MED 26020649 PMC4447368 10.1371/journal.pgen.1005207
+#>                                                                                                                                                                                                    title
+#> 1                                                                                                            Hybrid Dysgenesis in Drosophila simulans Associated with a Rapid Invasion of the P-Element.
+#> 2                                                                                        Chromosomal-Level Assembly of the Asian Seabass Genome Using Long Sequence Reads and Multi-layered Scaffolding.
+#> 3                                                 An Indel Polymorphism in the MtnA 3' Untranslated Region Is Associated with Gene Expression Variation and Local Adaptation in Drosophila melanogaster.
+#> 4 Germline Defects Caused by Smed-boule RNA-Interference Reveal That Egg Capsule Deposition Occurs Independently of Fertilization, Ovulation, Mating, or the Presence of Gametes in Planarian Flatworms.
+#> 5                                                                                                                                                    Virus Satellites Drive Viral Evolution and Ecology.
+#> 6                                                                      A Simple Auxin Transcriptional Response System Regulates Multiple Morphogenetic Processes in the Liverwort Marchantia polymorpha.
 #>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             authorString
 #> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Hill T, Schlötterer C, Betancourt AJ.
[…]
