### 2016-05-18 a0a5383
align with cran repository policy
