+++ vignettes/rebi-vignettes.Rmd
@@ -32,7 +32,7 @@ For more background, see:
 <https://europepmc.org/About>
 
 Europe PMC: a full-text literature database for the life sciences and platform 
-for innovation. (2014). Nucleic Acids Research, 43(D1), D1042–D1048. doi:[10.1093/nar/gku1061](http://doi.org/10.1093/nar/gku1061)
+for innovation. (2014). Nucleic Acids Research, 43(D1), D1042–D1048. <http://doi.org/10.1093/nar/gku1061>
 
 ## Installation
 
