+++ DESCRIPTION
@@ -1,23 +1,20 @@
 Package: rebi
+Title: R Interface to Europe PubMed Central RESTful Web Service
 Version: 0.0.999
-Maintainer: 'Najko Jahn' <najko.jahn@uni-bielefeld.de>
-Author: Najko Jahn
+Authors@R: person("Najko", "Jahn", email = "najko.jahn@gmail.com",
+                 role = c("aut", "cre"))
 License: GPL-3
-Title: rebi -- R Interface for Europe PMC RESTful Web Service
 URL: http://github.com/njahn82/rebi/
 BugReports: http://github.com/njahn82/rebi/issues
-Description: R Interface to Europe PMC RESTful Web Service. Europe PMC covers
-    life science literature and it gives access to open access full texts.
-    Coverage is not only restricted to Europe, but articles and abstracts are
-    indexed from all over the world. As a partner in the PMC International
-    (PMCi), Europe PMC ingests all PubMed content and extends its index with
-    other sources, including Agricola, a bibliographic database of citations to
-    the agricultural literature, or Biological Patents. In addition to
-    bibliographic data, rebi gives automated access to citations and references
-    that were indexed by Europe PMC. Links between life-science literature and
-    other EBI databases, including ENA, PDB or ChEMBL are also accessible.
-    External links provide a mechanism to link out to external providers such as
-    Wikipedia or research data repositories.
+Description: An R Client for the Europe PubMed Central RESTful Web Service. It
+    gives access to both metadata on life science literature and open access
+    full texts. Europe PMC indexes all PubMed content, and other literature
+    sources, including Agricola, a bibliographic database of citations to the
+    agricultural literature, or Biological Patents. In addition to bibliographic
+    metadata, rebi allows users to fetch citations and reference lists. Links
+    between life-science literature and other EBI databases, including ENA, PDB
+    or ChEMBL are also accessible. No registration or API key is required to use
+    this R client. See the vignettes for usage examples.
 LazyLoad: yes
 LazyData: yes
 VignetteBuilder: knitr
