### 2016-05-25 #29 @toph-allen
I'm so sorry I took even longer to respond to your changes—I had been budgeting more time to go through it than it actually took.

Reading through your responses to my suggestions, and looking through the accompanying changes to the code, I don't think I have any further issues.

Regarding the naming of your packages:

> Do you think `europepmc` as package name would be more appropriate since the EBI offers many other services that are not covered by this package?

I think that renaming the package `europepmc` or `epmc` would make its name more obviously connected with its main functions, and with the function prefixes. Those would be concrete benefits over the current name. It might be a bit of a hassle (or it might not be more than a "find and replace" and some touch-up)—and, you might just be fond of the current name. I'll leave that decision up to you.

Seems ok to me, though.
