### 2016-05-25 #29 @njahn82
Great, I would prefer to rename the package to `europepmc` as I have not initially submitted it to CRAN yet, but plan to do this within the next days.

Many thanks for reviewing this package!
