### 2016-05-25 #29 @noamross
Thanks, @toph-allen for reviewing and @njahn82 for keeping up this package!

@njahn82 I am getting one local error for a test that is skipped on Travis:

```
1. Failure (at test_epmc_citations.r#22): epmc_citations returns ---------------
epmc_citations("13814508") does not match 'No citing documents found'. Actual value: "Error in rebi_GET(path = path) : Not Found (HTTP 404).\n"
```

Once that's fixed, and you've made all the internal changes for changing the name, I think we're good to go.  You should be able to apply your changes to the package in the ropensci repo via a pull request.  I'm not sure about whether you have the ability to change _that_ repo's name.  @sckott: heads up, we may need to see that a name change populates through our systems properly.
