### 2016-05-25 #29 @njahn82
Thanks for letting me know. I found that Europe PMC API has started to give a 404 when no citations were found, which is not a nice behavior.

Compare
http://www.ebi.ac.uk/europepmc/webservices/rest/MED/13814508/citations
with
http://www.ebi.ac.uk/europepmc/webservices/rest/MED/9843981/citations

I will fix the error messages and corresponding tests.
