### 2016-05-26 0768e80
exclude message from test #6
