+++ README.md
@@ -1,19 +1,19 @@
-# rebi - R Interface to Europe PMC RESTful Web Service
+# europepmc - R Interface to Europe PMC RESTful Web Service
 
 
 
-[![Build Status](https://travis-ci.org/njahn82/rebi.svg?branch=master)](https://travis-ci.org/njahn82/rebi)
-[![Build status](https://ci.appveyor.com/api/projects/status/nm7ixfn2twuu1y45/branch/master?svg=true)](https://ci.appveyor.com/project/njahn82/rebi/branch/master)
-[![codecov.io](https://codecov.io/github/njahn82/rebi/coverage.svg?branch=master)](https://codecov.io/github/njahn82/rebi?branch=master)
+[![Build Status](https://travis-ci.org/njahn82/europepmc.svg?branch=master)](https://travis-ci.org/njahn82/europepmc)
+[![Build status](https://ci.appveyor.com/api/projects/status/nm7ixfn2twuu1y45/branch/master?svg=true)](https://ci.appveyor.com/project/njahn82/europepmc/branch/master)
+[![codecov.io](https://codecov.io/github/njahn82/europepmc/coverage.svg?branch=master)](https://codecov.io/github/njahn82/europepmc?branch=master)
 
-rebi facilitates access to [Europe PMC RESTful Web
+europepmc facilitates access to [Europe PMC RESTful Web
 Service](http://europepmc.org/RestfulWebService).
 
 [Europe PMC](http://europepmc.org/) covers life science literature and
-gives access to open access full texts. Europe 
-PMC ingests all PubMed content and extends its index with other sources, 
-including Agricola, a bibliographic database of citations to the agricultural 
-literature, or Biological Patents. 
+gives access to open access full texts. Europe
+PMC ingests all PubMed content and extends its index with other sources,
+including Agricola, a bibliographic database of citations to the agricultural
+literature, or Biological Patents.
 
 ![Index coverage](https://europepmc.org/wicket/resource/uk.bl.ukpmc.web.pages.faq.Help/images/EuropePMCContent-ver-4BB17F003F8F38DF2D3BBE48AB5896C6.png)
 
@@ -21,7 +21,7 @@ For more background, see:
 
 <https://europepmc.org/About>
 
-Europe PMC: a full-text literature database for the life sciences and platform 
+Europe PMC: a full-text literature database for the life sciences and platform
 for innovation. (2014). Nucleic Acids Research, 43(D1), D1042–D1048. doi:[10.1093/nar/gku1061](http://doi.org/10.1093/nar/gku1061)
 
 ## Installation
@@ -32,7 +32,7 @@ The latest development version can be installed using
 
 ```r
 require(devtools)
-install_github("ropensci/rebi")
+install_github("ropensci/europepmc")
 ```
 
 ## Search Europe PMC
@@ -44,12 +44,12 @@ PMC search syntax to `epmc_search()`.
 
 The search function helps to get a general overview about additional
 information types that are offered by Europe PMC and which can be retrieved
-through other `rebi`-functions. Columns inform whether open access full texts
+through other `europepmc`-functions. Columns inform whether open access full texts
 (`isOpenAccess`), cross-links to other EBI databases (`hasDbCrossReferences`),
 text-mined terms (`hasTextMinedTerms`) or references (`hasReferences`) are
 available.
 
-By default, `epmc_search` returns 25 records. To adjust the limit, simply use 
+By default, `epmc_search` returns 25 records. To adjust the limit, simply use
 the `limit` parameter.
 
 Either list of publication ids (`id_list = TRUE`) or key metadata
[…]
