+++ vignettes/rebi-vignettes.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Introducing rebi, an R interface to Europe PMC RESTful API"
+title: "Introducing europepmc, an R interface to Europe PMC RESTful API"
 author: "Najko Jahn"
 date: "13 May 2016"
 output: rmarkdown::html_vignette
@@ -20,10 +20,10 @@ knitr::opts_chunk$set(
 ```
 
 [Europe PMC](http://europepmc.org/) covers life science literature and
-gives access to open access full texts. Europe 
-PMC ingests all PubMed content and extends its index with other sources, 
-including Agricola, a bibliographic database of citations to the agricultural 
-literature, or Biological Patents. 
+gives access to open access full texts. Europe
+PMC ingests all PubMed content and extends its index with other sources,
+including Agricola, a bibliographic database of citations to the agricultural
+literature, or Biological Patents.
 
 ![Index coverage](https://europepmc.org/wicket/resource/uk.bl.ukpmc.web.pages.faq.Help/images/EuropePMCContent-ver-4BB17F003F8F38DF2D3BBE48AB5896C6.png)
 
@@ -31,7 +31,7 @@ For more background, see:
 
 <https://europepmc.org/About>
 
-Europe PMC: a full-text literature database for the life sciences and platform 
+Europe PMC: a full-text literature database for the life sciences and platform
 for innovation. (2014). Nucleic Acids Research, 43(D1), D1042–D1048. <http://doi.org/10.1093/nar/gku1061>
 
 ## Installation
@@ -42,7 +42,7 @@ The latest development version can be installed using
 
 ```r
 require(devtools)
-install_github("ropensci/rebi")
+install_github("ropensci/europepmc")
 ```
 
 ## Search Europe PMC
@@ -54,12 +54,12 @@ PMC search syntax to `epmc_search()`.
 
 The search function helps to get a general overview about additional
 information types that are offered by Europe PMC and which can be retrieved
-through other `rebi`-functions. Columns inform whether open access full texts
+through other `europepmc`-functions. Columns inform whether open access full texts
 (`isOpenAccess`), cross-links to other EBI databases (`hasDbCrossReferences`),
 text-mined terms (`hasTextMinedTerms`) or references (`hasReferences`) are
 available.
 
-By default, `epmc_search` returns 25 records. To adjust the limit, simply use 
+By default, `epmc_search` returns 25 records. To adjust the limit, simply use
 the `limit` parameter.
 
 Either list of publication ids (`id_list = TRUE`) or key metadata
@@ -81,7 +81,7 @@ my_data <- epmc_search(query = 'ISSN:1553-7404 HAS_EMBL:y')
 head(my_data)
 ```
 
-### Get results number 
+### Get results number
 
 Count hits before with `epmc_hits` to define limit. For example, get list of ids
 that represent articles referencing DataCite DOIs:
@@ -90,7 +90,7 @@ that represent articles referencing DataCite DOIs:
 query <- "ACCESSION_TYPE:doi"
 epmc_hits(query)
 # set limit to 10 records
-my_data <- epmc_search(query = "ACCESSION_TYPE:doi", limit = 10, 
+my_data <- epmc_search(query = "ACCESSION_TYPE:doi", limit = 10,
                        id_list = TRUE)
 head(my_data)
 attr(my_data, "hit_count")
@@ -101,13 +101,13 @@ attr(my_data, "hit_count")
 Use [ORCID](http://orcid.org/) to search for personal publications:
 
 ```{r}
-my_data <- epmc_search(query = 'AUTHORID:"0000-0002-7635-3473"') 
+my_data <- epmc_search(query = 'AUTHORID:"0000-0002-7635-3473"')
[…]
