+++ man/europepmc.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/rebi.r
-\name{rebi}
-\alias{rebi}
-\title{rebi - R client for the Europe PMC RESTful article API}
+\name{europepmc}
+\alias{europepmc}
+\title{europepmc - R client for the Europe PMC RESTful article API}
 \description{
-What is rebi?
+What is europepmc?
 }
 \details{
-rebi facilitates access to Europe PMC RESTful Web Service. Europe
+europepmc facilitates access to Europe PMC RESTful Web Service. Europe
 PMC covers life science literature and gives access to open access full
 texts. Coverage is not only restricted to Europe, but articles and
 abstracts are indexed from all over the world. As a partner in the PMC
@@ -15,7 +15,7 @@ International (PMCi), Europe PMC ingests all PubMed content and extends its
 index with other sources, including Agricola, a bibliographic database of
 citations to the agricultural literature, or Biological Patents.
 
-Apart of searching abstracts and full text, rebi can be used to
+Apart of searching abstracts and full text, europepmc can be used to
 retrieve reference sections and citations, text-mined terms or cross-links
 to other databases hosted by the European Bioinformatics Institute (EBI).
 

[…]
