### 2016-05-26 #29 @sckott
> we may need to see that a name change populates through our systems properly.

okay, not sure how to handle this when there's a fork and original version - I'll look into this
