### 2016-05-31 3e48c8f
fix links to ci services
