+++ README.md
@@ -2,9 +2,9 @@
 
 
 
-[![Build Status](https://travis-ci.org/njahn82/rebi.svg?branch=master)](https://travis-ci.org/njahn82/rebi)
-[![Build status](https://ci.appveyor.com/api/projects/status/nm7ixfn2twuu1y45/branch/master?svg=true)](https://ci.appveyor.com/project/njahn82/rebi/branch/master)
-[![codecov.io](https://codecov.io/github/njahn82/rebi/coverage.svg?branch=master)](https://codecov.io/github/njahn82/rebi?branch=master)
+[![Build Status](https://travis-ci.org/ropensci/rebi.svg?branch=master)](https://travis-ci.org/ropensci/rebi)
+[![Build status](https://ci.appveyor.com/api/projects/status/nm7ixfn2twuu1y45/branch/master?svg=true)](https://ci.appveyor.com/project/ropensci/rebi/branch/master)
+[![codecov.io](https://codecov.io/github/ropensci/rebi/coverage.svg?branch=master)](https://codecov.io/github/ropensci/rebi?branch=master)
 
 rebi facilitates access to [Europe PMC RESTful Web
 Service](http://europepmc.org/RestfulWebService).
@@ -64,61 +64,61 @@ my_data <- epmc_search(query = 'Gabi-Kat')
 # first six records
 head(my_data)
 #>         id source     pmid      pmcid                          doi
-#> 1 26842807    MED 26842807 PMC4740857            10.1038/srep20309
-#> 2 26930070    MED 26930070 PMC4773003 10.1371/journal.pone.0150254
-#> 3 26824478    MED 26824478 PMC4733102 10.1371/journal.pone.0148335
-#> 4 27064270    MED 27064270 PMC4814454      10.3389/fpls.2016.00405
-#> 5 27015278    MED 27015278 PMC4807768 10.1371/journal.pgen.1005959
-#> 6 26343971    MED 26343971       <NA>   10.1016/j.molp.2015.08.011
+#> 1 27117628    MED 27117628 PMC4846993            10.1038/srep24971
+#> 2 26842807    MED 26842807 PMC4740857            10.1038/srep20309
+#> 3 26930070    MED 26930070 PMC4773003 10.1371/journal.pone.0150254
+#> 4 26957563    MED 26957563 PMC4861014           10.1093/jxb/erw096
+#> 5 27064270    MED 27064270 PMC4814454      10.3389/fpls.2016.00405
+#> 6 26824478    MED 26824478 PMC4733102 10.1371/journal.pone.0148335
 #>                                                                                                                                                                                                                  title
-#> 1                                                                                  Precocious leaf senescence by functional loss of PROTEIN S-ACYL TRANSFERASE14 involves the NPR1-dependent salicylic acid signaling.
-#> 2 The Arabidopsis Domain of Unknown Function 1218 (DUF1218) Containing Proteins, MODIFYING WALL LIGNIN-1 and 2 (At1g31720/MWL-1 and At4g19370/MWL-2) Function Redundantly to Alter Secondary Cell Wall Lignin Content.
-#> 3                                                                                            The Early-Acting Peroxin PEX19 Is Redundantly Encoded, Farnesylated, and Essential for Viability in Arabidopsis thaliana.
-#> 4                                                                                          Photosystem II Repair and Plant Immunity: Lessons Learned from Arabidopsis Mutant Lacking the THYLAKOID LUMEN PROTEIN 18.3.
-#> 5                                                                                                    MicroProtein-Mediated Recruitment of CONSTANS into a TOPLESS Trimeric Complex Represses Flowering in Arabidopsis.
[…]
