### 2016-05-31 #29 @sckott
name changed, https://github.com/ropensci/europepmc
