### 2016-05-31 05a6b7e
fixing some other renameing issues
