+++ vignettes/rebi-vignettes.Rmd
@@ -68,7 +68,7 @@ information  (`id_list = FALSE`, default option) are returned.
 For instance, search for abstracts and full texts that mention `Gabi-Kat`:
 
 ```{r}
-library(rebi)
+library(europepmc)
 my_data <- epmc_search(query = 'Gabi-Kat')
 # first six records
 head(my_data)
