### 2016-05-31 #29 @njahn82
Great, all references to the former package name should now be changed to `europepmc`.

Many thanks for all your comments and suggested improvements, I really appreciate your help.
